fix: round-8 archaeology — P-1 empty array + P-6 DNS hostnames + P-8 DBF_CHAR signed
These are direct sibling fixes to long-standing C-reference bugs surfaced by a
30-year EPICS C history audit (BUG_ARCHAEOLOGY.md §3 P-1, P-6, P-8).

P-1 (libca 8cc20393f / a7bf59079 / a42197f0d, DB 12cfd418d):
- types/value.rs::from_bytes_array previously short-circuited on
  `count <= 1`, treating an empty array as a scalar decode. On GET
  this produced CaError::Protocol("char data empty" / "...too small")
  for every empty waveform read; on WRITE the server silently
  degraded a count=0 array PUT to a one-element scalar PUT.
- Now: count=0 dispatches to the typed empty-array variant
  (StringArray, ShortArray, ...). count=1 still falls through to
  the scalar decoder (legitimate "scalar shaped as array of one").

P-6 (libca a8e8d22c3):
- config/env.rs::parse_addr_list_with_port previously accepted only
  literal IP entries and silently dropped DNS hostnames. The C
  reference had a 32-byte hostname truncation bug; we had a stricter
  "drop the whole token" bug — same operator-visible "Empty PV
  search address list" symptom with no actionable error.
- Now: DNS hostnames resolve via ToSocketAddrs. Plain `host`,
  `host:port`, `[ipv6]:port`, bare `ip` and `ip:port` all parse;
  resolution failures log at debug instead of silently dropping.

P-8 (libca 7cb80d5a1):
- types/codec.rs encode_units_limits_u8 + DBF_CHAR decode used `u8`
  end-to-end. Per the C spec DBF_CHAR is `epicsInt8` (signed).
  Result: a peer C IOC sending CTRL_CHAR with `lower_disp_limit =
  -10` puts byte 0xF6 on the wire; epics-rs read it as 246.0
  (`data[off] as f64`). Symmetric write path saturated -10.0_f64
  to 0u8 silently. Negative DRVL/LOPR/HIHI on DBF_CHAR records
  lost sign on every CA hop.
- Now: encode via `(*l as i8) as u8` (saturating), decode via
  `(data[off] as i8) as f64`. Storage stays `EpicsValue::Char(u8)`
  (API-stable); only the wire boundary's sign-aware re-cast.

Tests added (5 new):
- p1_empty_array_double_decodes_as_empty_doublearray
- p1_count_one_double_falls_through_to_scalar
- p1_empty_array_string_decodes_as_empty_stringarray (40-byte slot)
- p1_empty_array_all_dbr_types_round_trip (every variant + encode/decode round-trip)
- test_encode_gr_char_saturates_out_of_range (P-8 saturation)
- test_encode_gr_char_with_metadata: rewrote to assert signed
  semantics (limits 100/-10 instead of 255/0).

Tests: 2983/2983 passing; clippy --workspace clean.

+++ crates/epics-base-rs/src/types/codec.rs
@@ -380,7 +380,12 @@ fn encode_units_limits_i32(
     }
 }
 
-/// units(8) + n limits as u8
+/// units(8) + n limits as u8 (DBF_CHAR limits, transmitted as
+/// SIGNED `epicsInt8` per the libca spec — `7cb80d5a1`). Encoding
+/// path: `f64 limit → i8 (saturating) → u8 bit-pattern → wire byte`.
+/// Without the i8 intermediate `(-10.0_f64) as u8` saturates to 0
+/// in Rust, silently destroying negative DRVL/LOPR/HOPR/HIHI on
+/// DBF_CHAR records. P-8 finding (BUG_ARCHAEOLOGY).
 fn encode_units_limits_u8(
     buf: &mut Vec<u8>,
     snapshot: &crate::server::snapshot::Snapshot,
@@ -389,7 +394,9 @@ fn encode_units_limits_u8(
     encode_units(buf, snapshot);
     let limits = get_limits(snapshot, n_limits);
     for l in &limits[..n_limits] {
-        buf.push(*l as u8);
+        // Saturating cast f64 → i8 → wire byte. `as i8` on f64 in
+        // Rust ≥1.45 saturates to [i8::MIN, i8::MAX] (no UB).
+        buf.push((*l as i8) as u8);
     }
 }
 
@@ -688,7 +695,12 @@ fn decode_gr_ctrl(
             let mut limits = [0.0f64; 8];
             for i in 0..n_limits {
                 if off < data.len() {
-                    limits[i] = data[off] as f64;
+                    // Decode as SIGNED i8 → f64 (DBF_CHAR is
+                    // epicsInt8 per libca 7cb80d5a1). The previous
+                    // `data[off] as f64` route read 0xF6 as 246.0
+                    // instead of -10.0, silently flipping signs on
+                    // CTRL_CHAR / GR_CHAR negative DRVL/LOPR limits.
+                    limits[i] = (data[off] as i8) as f64;
                 }
                 off += 1;
             }
